requirements.txt
Django==1.3.1
django-dajax==0.9.2
django-dajaxice==0.7
django-online-status==0.1.0
django-thumbs==0.4
html5lib==0.999
httpagentparser==1.7.6
Pillow==2.8.1
pisa==3.0.33
PyPDF2==1.24
python-magic==0.4.6
reportlab==2.7
requests==2.7.0
six==1.9.0
